added yes-no prompt

+++ magefiles/users.go
@@ -9,18 +9,29 @@ import (
 	"github.com/magefile/mage/mg"
 	k8sRoles "github.com/perithompson/awesome-k8s-mage/pkg/k8s/roles"
 	k8sUser "github.com/perithompson/awesome-k8s-mage/pkg/k8s/users"
+	"github.com/perithompson/awesome-k8s-mage/pkg/prompt"
 )
 
 type User mg.Namespace
 
 // Create creates a user
 func (User) Create(username, namespace string) error {
-	return k8sUser.CreateServiceAccount(username, namespace)
+	ok := prompt.YesNo(fmt.Sprintf("Are you sure you want to create %s/%s", namespace, username), true)
+	if ok {
+		return k8sUser.CreateServiceAccount(username, namespace)
+	} else {
+		return nil
+	}
 }
 
 // Delete deletes a user
 func (User) Delete(username, namespace string) error {
-	return k8sUser.DeleteServiceAccount(username, namespace)
+	ok := prompt.YesNo(fmt.Sprintf("Are you sure you want to delete %s/%s", namespace, username), false)
+	if ok {
+		return k8sUser.DeleteServiceAccount(username, namespace)
+	} else {
+		return nil
+	}
 }
 
 // CanThey runs can-i as the users kubeconfig
